[WCBRD-1336] Updates from Clibv1.5 : source code files

+++ src/comps/mcuxClMath/src/mcuxClMath_QDash.c
@@ -1,5 +1,5 @@
 /*--------------------------------------------------------------------------*/
-/* Copyright 2020-2022 NXP                                                  */
+/* Copyright 2020-2023 NXP                                                  */
 /*                                                                          */
 /* NXP Confidential. This software is owned or controlled by NXP and may    */
 /* only be used strictly in accordance with the applicable license terms.   */
@@ -17,9 +17,9 @@
  */
 
 
-#include <stdint.h>
-#include <mcuxCsslFlowProtection.h>
+#include <mcuxClCore_Platform.h>
 #include <mcuxClCore_FunctionIdentifiers.h>
+#include <mcuxCsslFlowProtection.h>
 
 #include <mcuxClPkc.h>
 #include <mcuxClMath_Functions.h>
@@ -116,7 +116,7 @@ MCUX_CSSL_FP_PROTECTED_TYPE(void) mcuxClMath_QSquared(uint32_t iQSqr_iNShifted_i
     MCUX_CSSL_FP_FUNCTION_ENTRY(mcuxClMath_QSquared,
         MCUX_CSSL_FP_FUNCTION_CALLED(mcuxClMath_QDash));
 
-    uint16_t len = (uint16_t) MCUXCLPKC_PS1_GETOPLEN();
+    uint16_t len = MCUXCLPKC_PS1_GETOPLEN();
     MCUX_CSSL_FP_FUNCTION_CALL_VOID(mcuxClMath_QDash(iQSqr_iNShifted_iN_iT, len));
 
     MCUX_CSSL_FP_FUNCTION_EXIT_VOID(mcuxClMath_QSquared);
